Handle error + Java Bean Validation + Table constraint

+++ src/main/java/com/myproject/student/student/StudentService.java
@@ -1,6 +1,8 @@
 package com.myproject.student.student;
 
 
+import com.myproject.student.student.exception.BadRequestException;
+import com.myproject.student.student.exception.StudentNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -16,12 +18,20 @@ public class StudentService {
     }
 
     public void addStudent(Student student) {
-        //check if email is taken
+        Boolean existsEmail = studentRepository
+                .selectExistsEmail(student.getEmail());
+        if (existsEmail) {
+            throw new BadRequestException(
+                    "Email " + student.getEmail() + " taken");
+        }
         studentRepository.save(student);
     }
 
     public void deleteStudent(Long studentId) {
-        // check if student exists
+        if(!studentRepository.existsById(studentId)) {
+            throw new StudentNotFoundException(
+                    "Student with id " + studentId + " does not exists");
+        }
         studentRepository.deleteById(studentId);
     }
 }
